Fixed result data type of all fixtures to long double so all of them return data in compatible format

+++ fixture_runner.h
@@ -22,11 +22,11 @@ private:
     static void CreateTrivialFixtures( std::vector<std::shared_ptr<Fixture>>& fixtures );
     static void CreateDampedWave2DFixtures( 
         std::vector<std::shared_ptr<Fixture>>& fixtures,
-        std::vector<std::shared_ptr<FixtureThatReturnsData<cl_float>>>& fixturesWithData );
+        std::vector<std::shared_ptr<FixtureThatReturnsData<long double>>>& fixturesWithData );
     static void CreateKochCurveFixtures( 
         std::vector<std::shared_ptr<Fixture>>& fixtures,
-        std::vector<std::shared_ptr<FixtureThatReturnsData<cl_float>>>& fixturesWithData,
-        std::vector<std::shared_ptr<FixtureThatReturnsData<cl_float>>>& fixtureToWriteResultToSVG );
+        std::vector<std::shared_ptr<FixtureThatReturnsData<long double>>>& fixturesWithData,
+        std::vector<std::shared_ptr<FixtureThatReturnsData<long double>>>& fixtureToWriteResultToSVG );
 
     static std::vector<boost::compute::device> FillDevicesList();
     static std::unordered_map<cl_device_id, boost::compute::context> FillContextsMap(
